Move function closer to caller
