+++ clinical/src/main/kotlin/com/hartwig/actin/clinical/curation/config/CurationConfigValidation.kt
@@ -161,33 +161,6 @@ inline fun <reified T : Enum<T>> validateEnum(
     }
 }
 
-fun validateIgnore(
-    curationCategory: CurationCategory,
-    input: String,
-    fieldName: String,
-    ignore: Boolean,
-    icdCodes: Set<IcdCode>,
-    year: Int?,
-    month: Int?,
-): List<CurationConfigValidationError> {
-    val anyFieldNotEmpty = icdCodes.isNotEmpty() || year != null || month != null
-    return if (ignore && anyFieldNotEmpty) {
-        listOf(
-            CurationConfigValidationError(
-                curationCategory,
-                input,
-                fieldName,
-                "<ignore>",
-                "not <ignore>",
-                "Cannot specify ICD codes, year, or month when ignore is true"
-            )
-        )
-    } else {
-        emptyList()
-    }
-}
-
-
 private fun String.toValidatedBoolean(): Boolean? {
     return if (this == "1") {
         true
